requirements.txt
python-gnupg==0.5.0
Flask==2.2.3
google-cloud-pubsub==2.16.0
google-cloud-secret-manager==2.16.1
google-cloud-storage==2.8.0